Update README.md
# SpaceProjectXL

[GamePlay Link](https://ewdlop.itch.io/spaceprojectxldemo)

## Introduction
SpaceProjectXL is an exciting space shooter game where players navigate through space, battling enemies and avoiding obstacles. The game features various levels, power-ups, and challenging bosses.

## Installation
To set up the development environment and install dependencies, follow these steps:
1. Clone the repository: `git clone https://github.com/ewdlop/SpaceProjectXL.git`
2. Navigate to the project directory: `cd SpaceProjectXL`
3. Open the project in Unity: `unity .`
4. Install any required packages through the Unity Package Manager.

## Usage
To run the project, open the project in Unity and press the Play button in the Unity Editor. Use the arrow keys to navigate the spaceship and the spacebar to shoot. Avoid obstacles and defeat enemies to progress through the levels.

## Contributing
We welcome contributions to the project! To contribute, follow these guidelines:
1. Fork the repository.
2. Create a new branch for your feature or bugfix: `git checkout -b my-feature-branch`
3. Make your changes and commit them: `git commit -m 'Add new feature'`
4. Push to the branch: `git push origin my-feature-branch`
5. Create a pull request detailing your changes.

## License
This project is licensed under the MIT License. See the [LICENSE](LICENSE) file for more information.

## Gameplay ScreenShots

![image](https://github.com/ewdlop/SpaceProjectXL/assets/25368970/e4adfc79-1f95-4858-b7ee-f5fdb3266b83)
![image](https://github.com/ewdlop/SpaceProjectXL/assets/25368970/892110dc-e2af-4346-9a9d-10ab3516c9ae)
![image](https://github.com/ewdlop/SpaceProjectXL/assets/25368970/596761b5-30e7-4aaf-8640-c87151a028b2)
![image](https://github.com/ewdlop/SpaceProjectXL/assets/25368970/07382526-753b-4acf-9eaf-63f6a556e6d0)
![image](https://github.com/ewdlop/SpaceProjectXL/assets/25368970/311022c3-89d0-4ee6-8589-90e0e0deed7e)